requirements.txt
Flask==3.0.0
streamlit==1.30.0
requests==2.31.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.2
pandas==2.1.4
PyPDF2==3.0.1
python-pptx==0.6.21
py2neo==2021.2.3
pycountry==22.3.5
tqdm==4.66.1